chore: move hooks about
