+++ react-components/src/components/RevealToolbar/LayersContainer/LayersContainer.tsx
@@ -10,7 +10,7 @@ import { PointCloudLayersContainer } from './PointCloudLayersContainer';
 import { useState, type ReactElement, useEffect, type MouseEvent } from 'react';
 import { type Reveal3DResourcesLayerStates, type Reveal3DResourcesLayersProps } from './types';
 import { useReveal } from '../../RevealContainer/RevealContext';
-import { useLayersUrlParams } from '../../../hooks/useUrlStateParam';
+import { useLayersUrlParams } from '../hooks/useUrlStateParam';
 
 const LayersContainer = ({ props }: { props: Reveal3DResourcesLayersProps }): ReactElement => {
   const viewer = useReveal();
