+++ react-components/src/components/RevealToolbar/MeasurementButton.tsx
@@ -8,7 +8,10 @@ import { Button, Tooltip as CogsTooltip } from '@cognite/cogs.js';
 import { type Measurement } from '@cognite/reveal/tools';
 import { FEET_TO_INCHES, METERS_TO_FEET } from '../../utilities/constants';
 import { useTranslation } from '../i18n/I18n';
-import { useAddMeasurementsToUrl, useInitializedMeasurementTool } from './measurementUrlStateHooks';
+import {
+  useAddMeasurementsToUrl,
+  useInitializedMeasurementTool
+} from './hooks/measurementUrlStateHooks';
 
 export const distancesInFeetAndMeters = (distanceInMeters: number): string => {
   const distanceInFeet = distanceInMeters * METERS_TO_FEET;
